+++ src/D365FO.Cli/Commands/Find/FindCommands.cs
@@ -2,6 +2,7 @@
 using D365FO.Core.Analysis;
 using Spectre.Console.Cli;
 using D365FO.Cli.Commands.Get;
+using D365FO.Core.Bridge;
 
 namespace D365FO.Cli.Commands.Find;