+++ src/D365FO.Cli/Commands/Validate/ValidateMetadataCommand.cs
@@ -1,6 +1,7 @@
-using D365FO.Cli.Commands.Get;
+﻿using D365FO.Cli.Commands.Get;
 using D365FO.Core;
 using Spectre.Console.Cli;
+using D365FO.Core.Bridge;
 
 namespace D365FO.Cli.Commands.Validate;